# yuanzi-backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "小园子母婴记录应用后端 API (基于 PRD-V1.1)",
        "title": "Yuanzi API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/photo": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "分页获取指定宝宝的照片列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "照片"
                ],
                "summary": "获取照片列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "宝宝ID",
                        "name": "baby_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/model.ListResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/handler.PhotoResponse"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/photo/callback": {
            "post": {
                "description": "OSS 上传完成后的回调处理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "照片"
                ],
                "summary": "照片上传回调",
                "parameters": [
                    {
                        "description": "回调参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PhotoCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/photo/upload-url": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取阿里云 OSS 直传签名 URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "照片"
                ],
                "summary": "获取照片上传 URL",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PhotoUploadURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.PhotoUploadURLResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/photo/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定照片",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "照片"
                ],
                "summary": "删除照片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "照片ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.PhotoCallbackRequest": {
            "type": "object",
            "required": [
                "etag",
                "photo_id",
                "size"
            ],
            "properties": {
                "etag": {
                    "type": "string"
                },
                "photo_id": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "handler.PhotoResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "taken_at": {
                    "type": "string"
                },
                "thumb_url": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "handler.PhotoUploadURLRequest": {
            "type": "object",
            "required": [
                "baby_id",
                "content_type",
                "filename",
                "size"
            ],
            "properties": {
                "baby_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "content_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "filename": {
                    "type": "string",
                    "example": "photo.jpg"
                },
                "size": {
                    "type": "integer",
                    "example": 2048000
                }
            }
        },
        "handler.PhotoUploadURLResponse": {
            "type": "object",
            "properties": {
                "access_url": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "integer"
                },
                "form_data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "photo_id": {
                    "type": "string"
                },
                "thumb_url": {
                    "type": "string"
                },
                "upload_url": {
                    "type": "string"
                }
            }
        },
...
        "model.ListResponse": {
            "type": "object",
            "properties": {
                "list": {},
                "pagination": {
                    "$ref": "#/definitions/model.Pagination"
                }
            }
        },
        "model.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}